requirements.txt
fastapi
uvicorn
civitdl
pyyaml

# test
httpx
pytest